refactor utils dir
