+++ src/csvUtils/getTransactionValue.ts
