+++ src/transactionSummary/getProjectedIncome.ts
@@ -1,5 +1,5 @@
 import type { ProjectedPaycheck, CCTransaction } from "../types";
-import { getTransactionValue } from "../utils/getTransactionValue";
+import { getTransactionValue } from "../csvUtils";
 import { config } from "../config";
 import { getDaysInMonth } from "./getDaysInMonth";
