+++ src/transactionSummary/getUpdatedPaychecks.ts
@@ -1,5 +1,5 @@
 import type { ProjectedPaycheck, CCTransaction } from "../types";
-import { getTransactionValue } from "../utils/getTransactionValue";
+import { getTransactionValue } from "../csvUtils";
 
 // summary: dep
 export const getUpdatedPaychecks = (
